backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase
supabase==2.3.4
postgrest==0.15.0
gotrue==2.4.1

# ML & AI
scikit-learn==1.4.0
xgboost==2.0.3
transformers==4.37.0
sentence-transformers==2.3.1
torch==2.2.0
onnx==1.15.0
onnxruntime==1.17.0
faiss-cpu==1.7.4
scikit-surprise==1.1.3
shap==0.44.0
fairlearn==0.10.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1
aiofiles==23.2.1

# Caching & Rate Limiting
redis==5.0.1
slowapi==0.1.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
